+++ example/lib/loopback.dart
@@ -85,6 +85,11 @@ class _LoopBackState extends State<LoopBack> {
       });
     }
 
+
+
+                final b = recorder.isFilterActive(RecorderFilterType.autogain);
+                print('Is Filter Active: $b');
+
     /// Listen for microphne data.
     recorder.uint8ListStream.listen((chunks) {
       if (audioSource != null) {
@@ -149,6 +154,7 @@ class _LoopBackState extends State<LoopBack> {
       sampleRate: sampleRate,
       channels: recorderChannels,
     );
+
     recorder
       ..start()
       ..startStreamingData();
@@ -168,6 +174,13 @@ class _LoopBackState extends State<LoopBack> {
         Row(
           mainAxisSize: MainAxisSize.min,
           children: [
+            OutlinedButton(
+              onPressed: () {
+                final b = recorder.isFilterActive(RecorderFilterType.autogain);
+                print('Is Filter Active: $b');
+              },
+              child: const Text('Is Filter Active'),
+            ),
             OutlinedButton(
               onPressed: () {
                 init();
@@ -226,7 +239,7 @@ class _LoopBackState extends State<LoopBack> {
 
         // if (echoCancellation) EchoCancellationSliders(),
 
-        const Bars(),
+        // const Bars(),
       ],
     );
   }