+++ lib/flutter_recorder.dart
@@ -5,4 +5,5 @@ export 'src/audio_data_container.dart';
 export 'src/enums.dart' show PCMFormat, RecorderChannels;
 export 'src/filters/autogain.dart';
 export 'src/filters/echo_cancellation.dart';
+export 'src/filters/filters.dart' show RecorderFilterType;
 export 'src/flutter_recorder.dart';