+++ lib/src/filters/echo_cancellation.dart
@@ -21,7 +21,7 @@ enum EchoCancellationEnum {
 }
 
 abstract class _EchoCancellationInternal extends FilterBase {
-  const _EchoCancellationInternal() : super(FilterType.echoCancellation);
+  const _EchoCancellationInternal() : super(RecorderFilterType.echoCancellation);
 
   EchoCancellationEnum get queryEchoDelayMs => EchoCancellationEnum.echoDelayMs;
   EchoCancellationEnum get queryEchoAttenuation =>